requirements.txt
pandas>=2.0
numpy>=1.24
pydantic>=2.0
python-dotenv>=1.0
pytest>=8.0
ortools>=9.9
jupyter>=1.0
